Removed Reply interface and renamed BaseReply to Reply
